Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
PyJWT==2.9.0
python-dotenv==1.0.1
Werkzeug==3.0.3
Flask-Migrate==4.0.5     # optional but recommended for database migrations